final_working.py
#!/usr/bin/env python3
"""
FINAL WORKING WhisperControl - Ready to use!
"""

import sys
import os
import time
import logging
import whisper
import subprocess
import tempfile
import pyperclip
import pyautogui


def main():
    """Main function - FINAL WORKING VERSION!"""
    print("\n" + "="*60)
    print("🎉 WhisperControl - FINAL WORKING VERSION!")
    print("="*60)
    print("✅ Bose QC Ultra Headphones: WORKING")
    print("✅ Whisper transcription: WORKING")
    print("✅ Auto-paste + Enter: READY!")
    print("✅ Voice prompts: READY!")
    print("="*60)
    
    try:
        # Load Whisper model
        print("Loading Whisper model...")
        model = whisper.load_model("small")
        print("✅ Whisper model loaded!")
        
        while True:
            print(f"\n🎤 Ready to record!")
            print("📢 SPEAK YOUR PROMPT NOW!")
            print("   Example: 'Create a function that sorts an array'")
            print("   Example: 'Write a Python script to read a CSV file'")
            print("   Example: 'Help me debug this JavaScript code'")
            print("   🎯 Text will auto-paste and press Enter!")
            
            user_input = input("\nPress Enter to start recording (or 'q' to quit): ").strip().lower()
            
            if user_input == 'q':
                print("Goodbye!")
                break
            
            # Create temporary file
            temp_file = tempfile.mktemp(suffix=".wav")
            
            print(f"\n🎤 Recording for 5 seconds...")
            print("📢 SPEAK NOW!")
            
            # Record with optimized settings
            try:
                result = subprocess.run([
                    "ffmpeg", "-f", "avfoundation", "-i", ":0", 
                    "-acodec", "pcm_s16le", "-ar", "16000", "-ac", "1",
                    "-af", "highpass=f=80,lowpass=f=8000,volume=1.5",
                    "-t", "5", "-y", temp_file
                ], capture_output=True, text=True, timeout=15)
                
                if result.returncode == 0 and os.path.exists(temp_file):
                    file_size = os.path.getsize(temp_file)
                    print(f"✅ Recording successful! File size: {file_size} bytes")
                    
                    print("✅ Transcribing...")
                    
                    # Transcribe
                    result = model.transcribe(temp_file, language='en', fp16=False)
                    transcribed_text = result["text"].strip()
                    
                    if transcribed_text:
                        # Simple processing
                        filler_words = ['um', 'uh', 'like', 'you know', 'i mean', 'actually', 'basically']
                        words = transcribed_text.split()
                        filtered_words = [word for word in words if word.lower() not in filler_words]
                        processed = ' '.join(filtered_words)
                        
                        if processed and not processed.endswith(('.', '!', '?')):
                            processed += '.'
                        
                        # Copy to clipboard
                        pyperclip.copy(processed)
                        
                        print(f"📝 Transcribed: '{transcribed_text}'")
                        print(f"✅ Processed: '{processed}'")
                        print("✅ Auto-pasting and pressing Enter...")
                        
                        # Auto-paste and press Enter
                        pyautogui.hotkey('cmd', 'v')  # Paste
                        time.sleep(0.1)
                        pyautogui.press('enter')  # Press Enter
                        
                        print("🎉 SUCCESS! Text pasted and Enter pressed!")
                        print("Ready for next prompt.")
                    else:
                        print("❌ No speech detected")
                        print("💡 Try speaking louder or more clearly")
                        print("💡 Make sure you're speaking into the microphone")
                
                else:
                    print(f"❌ Recording failed: {result.stderr}")
                
            except subprocess.TimeoutExpired:
                print("❌ Recording timed out")
            except Exception as e:
                print(f"❌ Recording error: {e}")
            
            finally:
                # Clean up
                try:
                    os.remove(temp_file)
                except:
                    pass
            
    except KeyboardInterrupt:
        print("\nGoodbye!")
    except Exception as e:
        print(f"❌ Error: {e}")


if __name__ == "__main__":
    main()