requirements.txt
openai-whisper
sounddevice==0.4.6
soundfile==0.12.1
pynput==1.7.6
pyperclip==1.8.2
PyYAML==6.0.1
numpy==1.24.3
webrtcvad==2.0.10
scipy==1.11.4
psutil==5.9.6
